postdated: Change to 'txauthor.AuthoredTx' to 'AuthoredPostDatedTx' to include coincase tx
Coincase tx needed to be sent to btcd for processing, because it doesn't
exist in mempool and sending only the hash of it is not enough.
Implementation is not finished yet.

Fixes:
Change totalInput to correct value. TotalInput value of 'txauthor.AuthoredTx'
need to total amount of all inputs (not count of inputs).
Set locktime to created tx explicitly (don't set the coincase tx)